Move list of data folders to its own file
