+++ index.ts
@@ -9,23 +9,10 @@ import { fdir } from 'fdir';
 
 import { CompatData } from './types/types.js';
 import extend from './scripts/lib/extend.js';
+import dataFolders from './scripts/lib/data-folders.js';
 import { normalizePath, walk } from './utils/index.js';
 
 const dirname = fileURLToPath(new URL('.', import.meta.url));
-
-export const dataFolders = [
-  'api',
-  'browsers',
-  'css',
-  'html',
-  'http',
-  'javascript',
-  'mathml',
-  'svg',
-  'webassembly',
-  'webdriver',
-  'webextensions',
-];
 
 /**
  * Recursively load one or more directories passed as arguments.
