+++ scripts/lib/data-folders.ts
@@ -1,0 +1,17 @@
+/* This file is a part of @mdn/browser-compat-data
+ * See LICENSE file for more information. */
+
+export const dataFoldersMinusBrowsers = [
+  'api',
+  'css',
+  'html',
+  'http',
+  'javascript',
+  'mathml',
+  'svg',
+  'webassembly',
+  'webdriver',
+  'webextensions',
+];
+
+export default [...dataFoldersMinusBrowsers, 'browsers'];
